Runtime/RTAcpi: Some more work on the ASL -> AML compiler, can process our vbox-standard.dsl now, bugref:10733 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108031 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/acpi/acpi-compiler.cpp
@@ -683,7 +683,7 @@ static int rtAcpiTblAslParseTerminal(PRTACPIASLCU pThis, RTACPIASLTERMINAL enmTe
 {
     int rc = VINF_SUCCESS;
 
-    AssertReturn(enmTerminal > RTACPIASLTERMINAL_INVALID && enmTerminal < RT_ELEMENTS(g_aAslKeywords), VERR_INTERNAL_ERROR);
+    AssertReturn(enmTerminal > RTACPIASLTERMINAL_INVALID && (unsigned)enmTerminal < RT_ELEMENTS(g_aAslKeywords), VERR_INTERNAL_ERROR);
 
     *ppAstNd = NULL;
